pyproject.toml
[project]
name = "sheaf"
version = "0.1.0"
description = "Open source PDF hosting platform with local and Azure Blob storage"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "aiosqlite>=0.20,<1",
    "alembic>=1.14,<2",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3,<4",
    "passlib[bcrypt]>=1.7,<2",
    "pydantic-settings>=2.6,<3",
    "redis>=5.2,<6",
    "aiofiles>=24.1,<25",
]

[project.optional-dependencies]
azure = ["azure-storage-blob>=12.23,<13"]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "ruff>=0.8,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
